requirements.txt
fastapi
uvicorn
google-generativeai
pdfplumber
jinja2